LPS-189313 Move WorkflowMetricsIndex to portal-workflow-metrics-api to be reachable in portal-workflow-kaleo-metrics-integration
