+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-api/src/main/java/com/liferay/portal/workflow/metrics/search/index/WorkflowMetricsIndex.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.workflow.metrics.internal.search.index;
+package com.liferay.portal.workflow.metrics.search.index;
 
 import com.liferay.portal.kernel.exception.PortalException;
